created function to create age and solved iteration problem
