+++ scripts/countries/factory-functions.js
@@ -144,3 +144,120 @@ export const justLetters = (string) => {
 		.split(",")
 		.join("");
 };
+
+export const findAge = (arr) => {
+	let today = new Date();
+	let thisYear = today.getFullYear();
+	let thisMonth = today.getMonth() + 1;
+	let thisDay = today.getDate();
+	console.log(today, thisYear, thisMonth, thisDay);
+
+	arr.forEach((index) => {
+		let age;
+		let birth = index.birthday;
+		let birthArray = birth.split(",").join("").split("th").join("").split(" ");
+		let isMonth = /[A-Za-z]{3,}/;
+		let isYear = /[0-9]{3,4}/;
+		let isDay = /[A-za-z]{2,3}/;
+		let birthMonth;
+		let birthYear;
+		let birthDay;
+
+		if (isMonth.test(birthArray[0])) {
+			birthMonth = birthArray[0];
+			birthMonth.toLowerCase() === "january"
+				? (birthMonth = 1)
+				: birthMonth.toLowerCase() === "february"
+				? (birthMonth = 2)
+				: birthMonth.toLowerCase() === "march"
+				? (birthMonth = 3)
+				: birthMonth.toLowerCase() === "april"
+				? (birthMonth = 4)
+				: birthMonth.toLowerCase() === "may"
+				? (birthMonth = 5)
+				: birthMonth.toLowerCase() === "june"
+				? (birthMonth = 6)
+				: birthMonth.toLowerCase() === "july"
+				? (birthMonth = 7)
+				: birthMonth.toLowerCase() === "august"
+				? (birthMonth = 8)
+				: birthMonth.toLowerCase() === "september"
+				? (birthMonth = 9)
+				: birthMonth.toLowerCase() === "october"
+				? (birthMonth = 10)
+				: birthMonth.toLowerCase() === "november"
+				? (birthMonth = 11)
+				: birthMonth.toLowerCase() === "december"
+				? (birthMonth = 12)
+				: birthArray[0];
+		}
+		for (let i = 0; i < birthArray.length; i++) {
+			if (isYear.test(birthArray[i])) {
+				birthYear = birthArray[i];
+			}
+		}
+
+		if (!isDay.test(birthArray[1])) {
+			birthDay = birthArray[1];
+		}
+		let deathArray;
+		let deathYear;
+		let deathMonth;
+		let deathDay;
+		if (index.death !== "") {
+			let death = index.death;
+			deathArray = death.split(",").join("").split("th").join("").split(" ");
+			if (isMonth.test(deathArray[0])) {
+				deathMonth = deathArray[0];
+				deathMonth.toLowerCase() === "january"
+					? (deathMonth = 1)
+					: deathMonth.toLowerCase() === "february"
+					? (deathMonth = 2)
+					: deathMonth.toLowerCase() === "march"
+					? (deathMonth = 3)
+					: deathMonth.toLowerCase() === "april"
+					? (deathMonth = 4)
+					: deathMonth.toLowerCase() === "may"
+					? (deathMonth = 5)
+					: deathMonth.toLowerCase() === "june"
+					? (deathMonth = 6)
+					: deathMonth.toLowerCase() === "july"
+					? (deathMonth = 7)
+					: deathMonth.toLowerCase() === "august"
+					? (deathMonth = 8)
+					: deathMonth.toLowerCase() === "september"
+					? (deathMonth = 9)
+					: deathMonth.toLowerCase() === "october"
+					? (deathMonth = 10)
+					: deathMonth.toLowerCase() === "november"
+					? (deathMonth = 11)
+					: deathMonth.toLowerCase() === "december"
+					? (deathMonth = 12)
+					: deathArray[0];
+			}
+			for (let i = 0; i < deathArray.length; i++) {
+				if (isYear.test(deathArray[i])) {
+					deathYear = deathArray[i];
+				}
+			}
+			if (!isDay.test(deathArray[1])) {
+				deathDay = deathArray[1];
+			}
+
+			age = deathYear - birthYear - 1;
+			if (deathMonth > birthMonth) {
+				age++;
+			} else if (deathMonth == birthMonth && deathDay > birthDay) {
+				age++;
+			}
+		} else {
+			age = thisYear - birthYear - 1;
+			if (thisMonth > birthMonth) {
+				age++;
+			} else if (thisMonth == birthMonth && thisDay > birthDay) {
+				age++;
+			}
+		}
+		index.age = Math.abs(age);
+	});
+};
