+++ scripts/countries/leaders/leaders.js
@@ -2,7 +2,6 @@ import { makeLeader } from "../factory-functions.js";
 
 export let leaderCollector = [];
 
-
 makeLeader(
 	leaderCollector,
 	"Ukraine",
@@ -11,7 +10,7 @@ makeLeader(
 	"1st President of Ukraine",
 	"January 10th 1934",
 	"May 10th 2022",
-	"Social Democratic Party of Ukraine",
+	"Social Democratic Party of Ukraine"
 );
 
 makeLeader(
@@ -21,6 +20,7 @@ makeLeader(
 	"Volodymyr Zelenskyy",
 	"President of Ukraine",
 	"January 25th 1978",
+	"",
 	"Independent party"
 );
 
@@ -34,20 +34,20 @@ makeLeader(
 	"still active President",
 	"Catholic leftist politician",
 	"Second longest-serving president"
-	);
+);
 
 makeLeader(
 	leaderCollector,
 	"Japan",
 	"https://cloudfront-us-east-2.images.arcpublishing.com/reuters/RCIPB3FROBOYFPL7JTR6IMQRO4.jpg",
 	"Fumio Kishida",
-	"Prime Minister", 
-	"July 29, 1957", 
-	"", 
+	"Prime Minister",
+	"July 29, 1957",
+	"",
 	"Liberal Democratic Party (LDP)",
 	"Previously served as Minister for Foreign Affairs from 2012 to 2017",
 	"Chaired the LDP Policy Research Council from 2017 to 2020"
-  );
+);
 
 makeLeader(
 	leaderCollector,
